Add files via upload
This report aims to perform a data mining analysis on the business characteristics and text reviews from Yelp’s round 7 dataset challenge. 2486 restaurants from Las Vegas are selected as our instances for this research. The research is then divided into two parts. The first part is to explore the relationship between business characters and the ratings of the respective restaurants. The second part is to analyse the keywords from text reviews and its ability to predict the restaurants’ ratings. Three types of classifiers – OneR, C4.5 (J48) and Naïve Bayes are applied for the business dataset, while KNN (IBK), Jrip (RIPPER) and Multinomial Naïve Bayes are used for the review dataset. From our experiments, we discover that both datasets are able to provide us with insights on how the stars will change depending on the nature and value of the attributes.